backend/feat: insert `user__transactions`

+++ packages/backend/migrations/2025-08-13-080353_table_transaction_create/up.sql
@@ -4,9 +4,9 @@ CREATE TABLE user__transaction
     id                TEXT PRIMARY KEY,
     date_timestamp_ms INTEGER NOT NULL,
     description       TEXT    NOT NULL,
-    amount            REAL    NOT NULL,
+    amount            INTEGER NOT NULL,
     category_id       TEXT,
-    source_id         TEXT NOT NULL,
+    source_id         TEXT    NOT NULL,
     -- either "user" or "sepay"
 
     created_at        TIMESTAMP DEFAULT CURRENT_TIMESTAMP
